let summary writer create dir

+++ engine/engine.py
@@ -43,12 +43,10 @@ def get_args():
 
 def setup_output(output_dir, overwrite=False):
     output_dir = Path(output_dir)
-    output_dir.mkdir(parents=True, exist_ok=overwrite)
-    if utils.is_main_process():
-        if overwrite:
-            for child in output_dir.iterdir():
-                if child.is_file():
-                    child.unlink()
+    if utils.is_main_process() and output_dir.is_dir() and overwrite:
+        for child in output_dir.iterdir():
+            if child.is_file():
+                child.unlink()
 
     return output_dir
